Take away Gem2sService functions, we dont need a class
